requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
redis[hiredis]>=5.0.1
casbin>=1.40.0
argon2-cffi>=23.1.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
PyJWT>=2.8.0 